fe: new component and additional design changes
- Move renderIcon to new react component instead.
- Added design regarding max counter and colors/icons on notification bell.
